refactor : 객체 이동, 객체 이름 변경
