+++ src/main/java/core/di/factory/BeanScanner.java
@@ -1,4 +1,4 @@
-package core.mvc.tobe;
+package core.di.factory;
 
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
@@ -20,9 +20,9 @@ import java.util.*;
 import static core.util.ReflectionUtils.newInstance;
 import static java.util.Arrays.asList;
 
-public class ControllerScanner {
+public class BeanScanner {
 
-    private static final Logger logger = LoggerFactory.getLogger(core.mvc.tobe.ControllerScanner.class);
+    private static final Logger logger = LoggerFactory.getLogger(BeanScanner.class);
 
     private static final List<ArgumentResolver> argumentResolvers = asList(
                 new HttpRequestArgumentResolver(),
